requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
clear-cache==1.0.0
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
